♻️ refactor: keep diagnostics internal
