+++ internal/diagnostic/endpoint_test.go
@@ -1,27 +1,27 @@
-package endpoint
+package diagnostic
 
 import (
 	"strings"
 	"testing"
 )
 
-func TestForDiagnosticRemovesEndpointSecrets(t *testing.T) {
+func TestEndpointRemovesSecrets(t *testing.T) {
 	const raw = "https://user:canary-userinfo@collector.example:4318/v1/traces?api_key=canary-query#canary-fragment"
-	got := ForDiagnostic(raw)
+	got := Endpoint(raw)
 	want := "https://collector.example:4318/v1/traces"
 	if got != want {
-		t.Fatalf("ForDiagnostic(%q) = %q, want %q", raw, got, want)
+		t.Fatalf("Endpoint(%q) = %q, want %q", raw, got, want)
 	}
 	for _, secret := range []string{"canary-userinfo", "canary-query", "canary-fragment"} {
 		if strings.Contains(got, secret) {
-			t.Fatalf("ForDiagnostic leaked %q in %q", secret, got)
+			t.Fatalf("Endpoint leaked %q in %q", secret, got)
 		}
 	}
 }
 
-func TestForDiagnosticDoesNotEchoMalformedEndpoint(t *testing.T) {
+func TestEndpointDoesNotEchoMalformedValue(t *testing.T) {
 	const raw = "https://canary-userinfo@collector.example/%zz?api_key=canary-query#canary-fragment"
-	if got := ForDiagnostic(raw); got != invalid {
-		t.Fatalf("ForDiagnostic(%q) = %q, want %q", raw, got, invalid)
+	if got := Endpoint(raw); got != invalidEndpoint {
+		t.Fatalf("Endpoint(%q) = %q, want %q", raw, got, invalidEndpoint)
 	}
 }
